--- lib/ui/conponent/weather_view/weather_control_widgets.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WeatherControlWidgets extends StatelessWidget {
-  const WeatherControlWidgets({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: TextButton(
-            onPressed: () {},
-            child: const Text('Close'),
-          ),
-        ),
-        Expanded(
-          child: TextButton(
-            onPressed: () {},
-            child: const Text('Reload'),
-          ),
-        ),
-      ],
-    );
-  }
-}
