+++ lib/ui/view/weather_view/weather_view.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_training/ui/conponent/weather_item_widget.dart';
-import 'package:flutter_training/ui/conponent/weather_view/weather_control_widgets.dart';
 
 class WeatherView extends StatefulWidget {
   const WeatherView({super.key});
@@ -14,28 +13,50 @@ class _WeatherViewState extends State<WeatherView> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: SafeArea(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            const Spacer(),
-            const FractionallySizedBox(
-              widthFactor: 0.5,
-              child: WeatherTemperatureWidget(),
-            ),
-            Expanded(
-              child: Column(
-                children: const [
-                  SizedBox(height: 80),
-                  FractionallySizedBox(
-                    widthFactor: 0.5,
-                    child: WeatherControlWidgets(),
+        child: Center(
+          child: FractionallySizedBox(
+            widthFactor: 0.5,
+            child: Column(
+              children: [
+                const Spacer(),
+                const WeatherTemperatureWidget(),
+                Flexible(
+                  child: Column(
+                    children: const [
+                      SizedBox(height: 60),
+                      _Buttons(),
+                    ],
                   ),
-                ],
-              ),
+                ),
+              ],
             ),
-          ],
+          ),
         ),
       ),
+    );
+  }
+}
+
+class _Buttons extends StatelessWidget {
+  const _Buttons();
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Expanded(
+          child: TextButton(
+            onPressed: () {},
+            child: const Text('Close'),
+          ),
+        ),
+        Expanded(
+          child: TextButton(
+            onPressed: () {},
+            child: const Text('Reload'),
+          ),
+        ),
+      ],
     );
   }
 }
